Create lock line effect
